reworked structure, moved logic to serverside
