+++ .gitignore
@@ -1,0 +1,3 @@
+**/citydata/cities1000_BP.json
+**/citydata/cities15000.txt
+**/node_modules
