+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "city-quiz-multiplayer",
+  "version": "1.0.0",
+  "description": "Multiplayer version of https://cityquiz.io/quizzes/world",
+  "main": "src/server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Evankhell0/city-quiz-multiplayer.git"
+  },
+  "author": "Evankhell0",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Evankhell0/city-quiz-multiplayer/issues"
+  },
+  "homepage": "https://github.com/Evankhell0/city-quiz-multiplayer#readme",
+  "dependencies": {
+    "express": "^4.18.2",
+    "http": "^0.0.1-security",
+    "socket.io": "^4.7.2"
+  }
+}
