--- server/public/css/index.css
@@ -1,33 +1,0 @@
-h1 {
-    font-size: 50px;
-    text-align: center
-}
-
-body {
-    background-color: #c4c3c2;
-}
-
-button {
-    background-color: lightblue;
-    color: black;
-    font-size: 25px;
-    margin: 15px;
-    width: 200px;
-    height: 200px;
-}
-
-#sgpl {
-    position: absolute;
-    left: 50%;
-    top: 50%;
-    -webkit-transform: translate(-50%, -50%);
-    transform: translate(-50%, -50%);
-}
-
-#sg {
-    float: left;
-}
-
-#pl {
-    float: right;
-}
