--- server/public/css/singleplayer.css
@@ -1,23 +1,0 @@
-h1 {
-    font-size: 35px;
-    text-align: center
-}
-
-div.center {
-    text-align: center;
-    margin: auto;
-    width: 100%;
-}
-
-#Eingabe {
-    outline: 2px solid white;
-    border: 2px solid #000000;
-    height: 35px;
-    width: 375px;
-}
-
-#countrySelection {
-    border: 2px solid #000000;
-    height: 35px;
-    width: 150px;
-}
