--- server/public/index.html
@@ -1,17 +1,0 @@
-<html>
-
-<head>
-  <meta charset="utf-8"/>
-  <title>City Quiz</title>
-  <link rel="stylesheet" href="css/index.css">
-</head>
-
-<body>
-  <h1>City Quiz</h1>
-  <div id="sgpl">
-    <button onclick="window.location.href='Singleplayer.html'" id="sg" type="button" >Singleplayer</button>
-    <button onclick="window.location.href='Multiplayer.html'" id="pl" type="button">Multiplayer</button>
-  </div>
-</body>
-
-</html>
