--- server/public/singleplayer.html
@@ -1,20 +1,0 @@
-<html>
-
-<head>
-    <meta charset="utf-8"/>
-    <title>How Many World Cities can you name?</title>
-    <link rel="stylesheet" href="css/singleplayer.css">
-</head>
-
-<body>
-    <h1>How Many World Cities Can You Name?</h1>
-    <div class="center">
-        <input id="Eingabe" type="text" placeholder="" onsubmit="">
-        <select id="countrySelection"></select>
-    </div>
-
-    <script src="js/countrySelection.js"></script>
-    <script src="js/guessProcessing.js"></script>
-</body>
-
-</html>
