--- server/start.bat
@@ -1,3 +1,0 @@
-title City Quiz Multi
-node server.js
-pause
