--- server2/singleplayer.html
@@ -1,106 +1,0 @@
-<html>
-
-<head>
-    <meta charset="utf-8" />
-    <title>How Many World Cities can you name?</title>
-    <style>
-        h1 {
-            font-size: 35px;
-            text-align: center
-        }
-
-        div.center {
-            text-align: center;
-            margin: auto;
-            width: 100%;
-        }
-
-        #Eingabe {
-            outline: 2px solid white;
-            border: 2px solid #000000;
-            height: 35px;
-            width: 375px;
-        }
-
-        #countrySelection {
-            border: 2px solid #000000;
-            height: 35px;
-            width: 150px;
-        }
-    </style>
-</head>
-
-<body>
-    <h1>How Many World Cities Can You Name?</h1>
-    <div class="center">
-        <form id="form" action="">
-            <input id="Eingabe" autocomplete="off">
-            <select id="countrySelection">
-                <option value="Germany">Germany</option>
-            </select>
-            <button>Send</button>
-        </form>
-    </div>
-    <ul id="results"></ul>
-
-    <script src="/socket.io/socket.io.js"></script>
-
-    <script>
-        let guessedCities = [];
-
-        function processResult(result) {
-            var item = document.createElement('li');
-            if (result == false) {
-                // incorrect guess
-                item.textContent = "wrong guess";
-            } else {
-                // correct guess
-                if (checkIfGuessed(guessedCities, result)) {
-                    // => check if already guessed
-                    item.textContent = "already guessed (" + result.names[0] + ")";
-                } else {
-                    guessedCities.push(result);
-                    item.textContent = guessedCities.length + ". "  + result.names[0];
-                    socket.emit("correctGuess", result);
-                }
-            }
-            results.appendChild(item);
-        }
-
-        function checkIfGuessed(cityArray, city) {
-            for (let i = 0; i < cityArray.length; i++) {
-                if (cityArray[i].id == city.id) {
-                    return true;
-                }
-            }
-            return false;
-        }
-    </script>
-
-    <script>
-        var socket = io();
-
-        var form = document.getElementById("form");
-        var input = document.getElementById("Eingabe");
-        var select = document.getElementById("countrySelection");
-        var results = document.getElementById('results');
-
-        form.addEventListener("submit", function(e) {
-            e.preventDefault();
-            if (input.value) {
-                socket.emit("guess", input.value);
-                input.value = "";
-            }
-        });
-
-        socket.on("guessResult", function(result) {
-            processResult(result);
-        });
-
-        socket.on("newConnection", function(array) {
-            guessedCities = array;
-        });
-    </script>
-</body>
-
-</html>
