--- server2/start2.bat
@@ -1,3 +1,0 @@
-title City Quiz Multi
-node server2.js
-pause
