+++ start.bat
@@ -1,0 +1,3 @@
+title City Quiz Multi
+node .
+pause
